[Fix] #27 - apply code review
- you can control toast message height
- set defaultValue
- organize file (Utils)
